# specs/1-advanced-cloud-deployment/contracts/task-service-openapi.yaml
openapi: 3.0.3
info:
  title: Task Service API
  description: API for managing tasks with advanced features
  version: 1.0.0
servers:
  - url: http://task-service:8000
    description: Task Service running in Kubernetes

paths:
  /api/{user_id}/tasks:
    get:
      summary: List tasks with filtering, sorting, and search capabilities
      parameters:
        - name: user_id
          in: path
          required: true
          description: User identifier
          schema:
            type: string
        - name: status
          in: query
          description: Filter by task status
          schema:
            type: string
            enum: [pending, completed]
        - name: priority
          in: query
          description: Filter by task priority
          schema:
            type: string
            enum: [high, medium, low]
        - name: tag
          in: query
          description: Filter by task tag
          schema:
            type: string
        - name: created_after
          in: query
          description: Filter tasks created after this date
          schema:
            type: string
            format: date-time
        - name: created_before
          in: query
          description: Filter tasks created before this date
          schema:
            type: string
            format: date-time
        - name: sort_by
          in: query
          description: Field to sort by
          schema:
            type: string
            enum: [created_at, due_date, priority, title]
        - name: sort_order
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
        - name: q
          in: query
          description: Search query for full-text search
          schema:
            type: string
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      summary: Create a new task with advanced features
      parameters:
        - name: user_id
          in: path
          required: true
          description: User identifier
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid request data

  /api/{user_id}/tasks/search:
    get:
      summary: Search tasks by keyword
      parameters:
        - name: user_id
          in: path
          required: true
          description: User identifier
          schema:
            type: string
        - name: q
          in: query
          required: true
          description: Search query
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of results to return
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'

  /api/tasks/{task_id}:
    get:
      summary: Get a specific task
      parameters:
        - name: task_id
          in: path
          required: true
          description: Task identifier
          schema:
            type: integer
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
    put:
      summary: Update a task
      parameters:
        - name: task_id
          in: path
          required: true
          description: Task identifier
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
    delete:
      summary: Delete a task
      parameters:
        - name: task_id
          in: path
          required: true
          description: Task identifier
          schema:
            type: integer
      responses:
        '204':
          description: Task deleted successfully
        '404':
          description: Task not found

  /api/tasks/{task_id}/complete:
    put:
      summary: Mark a task as completed
      parameters:
        - name: task_id
          in: path
          required: true
          description: Task identifier
          schema:
            type: integer
      responses:
        '200':
          description: Task marked as completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found

components:
  schemas:
    Task:
      type: object
      required:
        - id
        - title
        - user_id
        - status
        - priority
        - created_at
      properties:
        id:
          type: integer
          description: Unique identifier for the task
        title:
          type: string
          description: Title of the task
        description:
          type: string
          description: Optional detailed description of the task
        status:
          type: string
          enum: [pending, completed]
          description: Status of the task
        priority:
          type: string
          enum: [high, medium, low]
          description: Priority level of the task
        tags:
          type: array
          items:
            type: string
          description: Array of tag strings for categorization
        due_date:
          type: string
          format: date-time
          description: Date and time when the task is due
        is_recurring:
          type: boolean
          description: Whether this task is recurring
        recurrence_settings:
          $ref: '#/components/schemas/RecurrenceSettings'
        reminder_settings:
          $ref: '#/components/schemas/ReminderSettings'
        user_id:
          type: string
          description: Identifier of the user who owns the task
        created_at:
          type: string
          format: date-time
          description: Timestamp when task was created
        updated_at:
          type: string
          format: date-time
          description: Timestamp when task was last updated
        completed_at:
          type: string
          format: date-time
          description: Timestamp when task was marked as completed

    RecurrenceSettings:
      type: object
      required:
        - pattern
        - interval
      properties:
        pattern:
          type: string
          enum: [daily, weekly, monthly, yearly]
          description: Recurrence pattern
        interval:
          type: integer
          description: How often to repeat (every N days/weeks/months/years)
        days:
          type: array
          items:
            type: string
            enum: [Mon, Tue, Wed, Thu, Fri, Sat, Sun]
          description: Days of week for weekly recurrence
        end_date:
          type: string
          format: date
          description: When to stop recurrence

    ReminderSettings:
      type: object
      required:
        - offset_minutes
        - channel
      properties:
        offset_minutes:
          type: integer
          description: Minutes before due time to send reminder
        channel:
          type: string
          enum: [in_app, email, push, sms]
          description: Notification channel

    CreateTaskRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
        priority:
          type: string
          enum: [high, medium, low]
          default: medium
        tags:
          type: array
          items:
            type: string
          maxItems: 10
        due_date:
          type: string
          format: date-time
        is_recurring:
          type: boolean
          default: false
        recurrence_settings:
          $ref: '#/components/schemas/RecurrenceSettings'
        reminder_settings:
          $ref: '#/components/schemas/ReminderSettings'

    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
        status:
          type: string
          enum: [pending, completed]
        priority:
          type: string
          enum: [high, medium, low]
        tags:
          type: array
          items:
            type: string
          maxItems: 10
        due_date:
          type: string
          format: date-time
        reminder_settings:
          $ref: '#/components/schemas/ReminderSettings'